money-test-a1b2c3/src/main/java/com/example/util/MoneyUtils.java
package com.example.util;

import java.math.BigDecimal;
import java.util.Arrays;

public class MoneyUtils {
    private static final int MW_GLIM = 5;
    private static final int MW_MCLIM = 50;
    private static final int MW_MLLIM = 6;

    private BigDecimal mwAmount;
    private int[] mwAmountGroup;
    private int mwAmountCents;
    private int mwWork;
    private char[] mwMoney;
    private int mwMcx;
    private String[] mwMoneyLines;
    private int mwMlx;
    private String mwWordText;
    private int mwWcx;
    private int mwGx;
    private int mwIx1;
    private int mwIx2;

    private static final String[] MW_GROUP_WORDS = {
        "TRILLION", "BILLION ", "MILLION ", "THOUSAND", "         "
    };

    private static final String[] MW_TENS_WORDS = {
        "TEN    ", "TWENTY ", "THIRTY ", "FORTY  ", "FIFTY  ",
        "SIXTY  ", "SEVENTY", "EIGHTY ", "NINETY "
    };

    private static final String[] MW_ONES_WORDS = {
        "ONE      ", "TWO      ", "THREE    ", "FOUR     ", "FIVE     ",
        "SIX      ", "SEVEN    ", "EIGHT    ", "NINE     ", "TEN      ",
        "ELEVEN   ", "TWELVE   ", "THIRTEEN ", "FOURTEEN ", "FIFTEEN  ",
        "SIXTEEN  ", "SEVENTEEN", "EIGHTEEN ", "NINETEEN "
    };

    public MoneyUtils() {
        this.mwAmount = BigDecimal.ZERO;
        this.mwAmountGroup = new int[MW_GLIM];
        this.mwAmountCents = 0;
        this.mwWork = 0;
        this.mwMoney = new char[MW_MCLIM];
        this.mwMcx = 0;
        this.mwMoneyLines = new String[MW_MLLIM];
        this.mwMlx = 0;
        this.mwWordText = "";
        this.mwWcx = 0;
        this.mwGx = 0;
        this.mwIx1 = 0;
        this.mwIx2 = 0;
    }

    public void setAmount(BigDecimal amount) {
        this.mwAmount = amount;
        updateAmountFields();
    }

    private void updateAmountFields() {
        BigDecimal dollars = mwAmount.setScale(0, BigDecimal.ROUND_DOWN);
        this.mwAmountCents = mwAmount.subtract(dollars).multiply(BigDecimal.valueOf(100)).intValue();

        String dollarString = dollars.toString();
        int length = dollarString.length();
        for (int i = 0; i < MW_GLIM; i++) {
            int start = Math.max(0, length - (i + 1) * 3);
            int end = length - i * 3;
            this.mwAmountGroup[MW_GLIM - 1 - i] = Integer.parseInt(dollarString.substring(start, end));
        }
    }

    public String convertToWords() {
        // Implementation of money conversion logic goes here
        // This would involve using the various arrays and fields to construct the word representation
        // of the monetary amount, similar to the COBOL logic
        return ""; // Placeholder return
    }

    // Additional helper methods as needed

    // Getter methods for fields if necessary
}